+++ nuxt.config.ts
@@ -7,7 +7,7 @@ export default defineNuxtConfig({
   compatibilityDate: "2025-05-15",
   devtools: { enabled: true },
   css: ["~/assets/css/main.css"],
-  modules: ["@nuxt/eslint", "@nuxt/icon", "@nuxtjs/color-mode"],
+  modules: ["@nuxt/eslint", "@nuxt/icon", "@nuxtjs/color-mode", "@pinia/nuxt"],
   vite: {
     plugins: [
       tailwindcss(),
@@ -20,5 +20,8 @@ export default defineNuxtConfig({
   },
   colorMode: {
     dataValue: "theme",
+  },
+  pinia: {
+    storesDirs: ["./app/stores/**"],
   },
 });
