+++ package.json
@@ -18,11 +18,13 @@
     "@nuxt/eslint": "1.4.1",
     "@nuxt/icon": "1.13.0",
     "@nuxtjs/color-mode": "3.5.2",
+    "@pinia/nuxt": "0.11.2",
     "@tailwindcss/vite": "^4.1.10",
     "better-auth": "^1.3.4",
     "drizzle-orm": "^0.44.2",
     "eslint": "^9.0.0",
     "nuxt": "^4.0.0",
+    "pinia": "^3.0.3",
     "tailwindcss": "^4.1.10",
     "vue": "^3.5.16",
     "vue-router": "^4.5.1",
